==> 02-data-wrangling.ipynb <==
# 2. Data Wrangling

- read data from the previous step
- select all "Falcon 9" launches and filter out the rest
- save new "Falcon 9" data to a `.csv` file


from pathlib import Path
import pandas as pd

import helpers as hlp

# Setup
pd.set_option("display.max_columns", None)
pd.set_option("display.max_colwidth", None)

INPUT_FILE = hlp.DATA_DIR / Path("01_collected_data.csv")
OUTPUT_FILE = hlp.DATA_DIR / Path("02_falcon9_data.csv")


## Read data


data_df = pd.read_csv(INPUT_FILE)
data_df.shape

data_df.columns

## Grab only Falcon 9 data


data_falcon9 = data_df.loc[data_df["booster_version"] == "Falcon 9"]
data_falcon9

## Clean data


data_falcon9.dtypes

data_falcon9.isnull().sum()
# We should see some null values on
# - the payload_mass column which we'll average out, and
# - the landing_pad column which we'll leave as is.

payloadmass_mean = data_falcon9["payload_mass"].mean()
data_falcon9 = data_falcon9.fillna({"payload_mass": payloadmass_mean})

data_falcon9["payload_mass"].isnull().sum()

## Save the clean data


data_falcon9.to_csv(OUTPUT_FILE, encoding="utf-8", mode="w", header=True, index=False)
