==> 04-eda-basic.ipynb <==
# 4. Basic Data Analysis

- adds an extra column to our dataset to classify successful and unsuccessful launches
- performs very basic numerical analysis on some of the columns


import pandas as pd
from pathlib import Path

import helpers as hlp

## Setup


pd.set_option("display.max_columns", None)
pd.set_option("display.max_colwidth", None)

INPUT_FILE = hlp.DATA_DIR / Path("02_falcon9_data.csv")  # data from step 2
OUTPUT_FILE = hlp.DATA_DIR / Path("04_eda_data.csv")


## Data Analysis


df = pd.read_csv(INPUT_FILE, encoding="utf-8", header=0, index_col=False)
df.astype(
    {
        "date": "datetime64[ns]",
        "booster_version": "object",
        "payload_mass": "float64",
        "orbit": "object",
        "launch_site": "object",
        "outcome": "object",
        "flights": "int64",
        "grid_fins": bool,
        "reused": bool,
        "legs": bool,
        "landing_pad": "object",
        "block": "float64",
        "reused_count": "int64",
        "serial": "object",
        "longitude": "object",
        "latitude": "object",
    }
)
df.head(5)

df.dtypes

df.shape

### create a new column as `landing_class`

distinct successful from unsuccessful launches:

- landing_class = 0 if bad_outcome
- landing_class = 1 otherwise


# Get a set of the possible bad outcomes
bad_outcomes = set(
    outcome
    for outcome in df["outcome"].unique()
    if "none" in outcome.lower() or "false" in outcome.lower()
)
bad_outcomes

landing_class = [0 if landing in bad_outcomes else 1 for landing in df["outcome"]]
len(landing_class)

df["class"] = landing_class

df.shape

## Simple Data Analysis


### % of missing values


df.isnull().sum() / len(df) * 100

### # of launches from each site


df["launch_site"].value_counts()

### mission outcome


# itemized launch outcomes
df["outcome"].value_counts()

### # of launches to each orbit


# number of launches to each orbit
df["orbit"].value_counts()

### obtain launch success rate


print(f"Success rate: {(sum(landing_class) / len(landing_class)) * 100:.2f}%")

# values should be the same, ~66.67%
print(f"Mean of class column: {df['class'].mean() * 100:.2f}%")

## Save dataframe


df.to_csv(OUTPUT_FILE, encoding="utf-8", mode="w", header=True, index=False)
