==> 05-eda-sql.ipynb <==
# 5. Exploratory Data Analysis (EDA)


import pandas as pd
import sqlite3
import prettytable
from pathlib import Path

import helpers as hlp


%load_ext sql

## Setup


pd.set_option("display.max_columns", None)
pd.set_option("display.max_colwidth", None)

INPUT_FILE = hlp.DATA_DIR / Path("spacex.csv")
DB_FILE = hlp.DATA_DIR / Path("05_spacex_data.sqlite3")

prettytable.DEFAULT = "DEFAULT"

## Open database


con = sqlite3.connect(DB_FILE)
cur = con.cursor()

## Save data to sqlite database


df = pd.read_csv(INPUT_FILE, encoding="utf-8", header=0, index_col=False)
df.to_sql("SPACEXTBL", con, if_exists="replace", index=False, method="multi")

## Connect to database


%sql sqlite:///data/05_spacex_data.sqlite3

## Run queries


# MISC. Get the name of all columns
%sql PRAGMA table_info(SPACEXTBL);

# Unique launch sites
%sql SELECT DISTINCT Launch_Site AS site FROM SPACEXTBL;

%%sql 
--# Launch sites that begin with CCA
SELECT Date, Booster_Version, Launch_Site, Orbit, Customer, Mission_Outcome, Landing_Outcome 
    FROM SPACEXTBL 
    WHERE Launch_Site LIKE 'CCA%' 
    LIMIT 5;

# Unique boosters
%sql SELECT DISTINCT Customer AS customer FROM SPACEXTBL WHERE customer LIKE 'NASA (CRS)%';

%sql SELECT DISTINCT Customer FROM SPACEXTBL;

%%sql 
--# Payload mass carried by boosters launched by NASA (CRS)
SELECT sum(PAYLOAD_MASS__KG_) AS nasa_payload_mass 
    FROM SPACEXTBL 
    WHERE customer LIKE 'NASA (CRS)%';

%sql SELECT DISTINCT Booster_Version FROM SPACEXTBL WHERE Booster_Version LIKE 'F9 v1.1%';

%%sql 
--# Average payload mass carried by booster version F9 v1.1
SELECT round(avg(PAYLOAD_MASS__KG_), 2) AS avg_f9_payload_mass 
    FROM SPACEXTBL 
    WHERE Booster_Version LIKE 'F9 v1.1%';

%sql SELECT DISTINCT Landing_outcome FROM SPACEXTBL;

%%sql 
--# first successful landing on a ground pad
SELECT min(Date) AS first_successful_landing 
    FROM SPACEXTBL 
    WHERE Landing_Outcome = 'Success (ground pad)';

%%sql
--# boosters successfully landed in dron ship and payload mass between 4000 and 6000
SELECT DISTINCT Booster_Version FROM SPACEXTBL
    WHERE 4000 < PAYLOAD_MASS__KG_ < 6000
    AND Landing_Outcome = 'Success (drone ship)';

%sql SELECT DISTINCT Mission_Outcome FROM SPACEXTBL;

%%sql
SELECT COUNT(Mission_Outcome) FROM SPACEXTBL
    WHERE Mission_Outcome = 'Success'
    OR Mission_Outcome LIKE 'Failure%';

%%sql
-- # total number of successful and unsuccessful mission outcomes
SELECT DISTINCT Mission_Outcome AS outcome, count(Mission_Outcome) AS count
    FROM SPACEXTBL
    GROUP BY Mission_Outcome;

%%sql 
-- # boosters that have carried the maximum payload mass
SELECT Booster_Version FROM SPACEXTBL
    WHERE PAYLOAD_MASS__KG_ = (SELECT max(PAYLOAD_MASS__KG_) 
    FROM SPACEXTBL);

%%sql
SELECT Date 
FROM SPACEXTBL
WHERE year(Date)='2015'
LIMIT 5

%%sql
-- #List the failed landing_outcomes in drone ship, their booster versions, and launch site names for in year 2015
SELECT substr(Date, 0, 5) AS year, 
Booster_Version AS booster, 
Launch_Site AS launch_site, 
Landing_Outcome AS outcome
    FROM SPACEXTBL 
    WHERE year = '2015' AND outcome = 'Failure (drone ship)';

%%sql
-- # month_names, failure landing_outcomes in drone ship, boosters, and launch_site for the months in year 2015
SELECT substr(date, 6, 2) As month, Booster_Version, Launch_Site FROM SPACEXTBL
WHERE Landing_Outcome LIKE '%Failure (drone ship)%'
AND CAST(substr(date, 0, 5) AS INTEGER) == 2015;

%%sql
--# count of landing outcomes between 2010-06-04 and 2017-03-20 in descending order
SELECT Landing_Outcome AS outcome, COUNT(Landing_Outcome) AS outcome_count FROM SPACEXTBL
GROUP BY outcome
ORDER BY outcome_count DESC;

%%sql
--# count of landing outcomes between 2010-06-04 and 2017-03-20 in descending order
SELECT Landing_Outcome AS outcome, COUNT(Landing_Outcome) AS outcome_count 
    FROM SPACEXTBL
    WHERE Date BETWEEN '2010-06-04' AND '2017-03-20'
    GROUP BY outcome
    ORDER BY outcome_count DESC;

df.columns

cur.close()
con.close()
