==> 09-dash-analysis.ipynb <==
# 9. Dash analysis

Run `python3 09-dashboard.py` fro your command line to start the dash server.

This Notebook brings some numbers in the form of tables to answer the following questions.

1. Launch site with a largest success launch rate and number of successes: **KSC LC-39A**
2. Launch site with a lowest success launch rate and number of failures: **CCAFS LC-40 76.92%**
3. Payload range with highest success rate: **2000 - 5000 kg**
4. Payload range with lowest success rate: **(0-2000 kg) & (5000-7000 kg)**
5. Booster with highest number of launch success: **F9 FT with 66.69%**


## Imports


import pandas as pd
from pathlib import Path

import helpers as hlp  # helper functions

## Setup


INPUT_FILE = hlp.DATA_DIR / Path("spacex_launch_dash.csv")
PAYLOAD_BINS = list(range(0, 10001, 2000))

## Load data


df = pd.read_csv(INPUT_FILE, encoding="utf-8", header=0, index_col=False)

# normalize column names
df.columns = [hlp.normalize_column_name(col) for col in df.columns]
df = df.drop("unnamed:_0", axis=1)
df.head()

## Common code


unique_sites = list(df["launch_site"].unique())
unique_booster_versions = df["booster_version"].apply(lambda x: x.split()[1]).unique()

## 1. Launch site with largest / lowest success launch rate


# intermediary lists
launches = []
succeeded = []
failed = []

# populate intermediary lists
for booster_version in unique_sites:
    launches.append(df.loc[(df["launch_site"] == booster_version)]["class"].count())
    succeeded.append(
        df.loc[(df["launch_site"] == booster_version) & (df["class"] == 1)][
            "class"
        ].count()
    )
    failed.append(
        df.loc[(df["launch_site"] == booster_version) & (df["class"] == 0)][
            "class"
        ].count()
    )

# create a new DataFrame to host launches, succeeded and failed launches
success_df = pd.DataFrame(
    data={
        "site": unique_sites,
        "launches": launches,
        "success": succeeded,
        "failure": failed,
    }
)
success_df["success_ratio"] = round(success_df["success"] / success_df["launches"], 4)
success_df["failure_ratio"] = round(success_df["failure"] / success_df["launches"], 4)
success_df

## 3. Payload success / failure rate


intervals_df = (
    pd.cut(df["payload_mass_kg"], PAYLOAD_BINS, right=True).value_counts().sort_index()
)
intervals_df = intervals_df.reset_index()
intervals_df

# intermediary lists
unique_intervals = []
launches = []
succeeded = []
failed = []
mins = []
maxs = []

# populate intermediary lists
for row in intervals_df.itertuples():
    _, interval, counts = row
    min_, max_ = [interval.left, interval.right]
    unique_intervals.append([min_, max_])
    mins.append(min_)
    maxs.append(max_)

    payload_condition = (min_ <= df["payload_mass_kg"]) & (df["payload_mass_kg"] < max_)
    launches.append(df.loc[payload_condition]["class"].count())
    succeeded.append(df.loc[(payload_condition) & (df["class"] == 1)]["class"].count())
    failed.append(df.loc[(payload_condition) & (df["class"] == 0)]["class"].count())

# create a new DataFrame to host intervals, succeeded and failed launches
success_df = pd.DataFrame(
    data={
        "interval": unique_intervals,
        "min": mins,
        "max": maxs,
        "launches": launches,
        "success": succeeded,
        "failure": failed,
    }
)
success_df["success_ratio"] = round(success_df["success"] / success_df["launches"], 4)
success_df["failure_ratio"] = round(success_df["failure"] / success_df["launches"], 4)
success_df = success_df.fillna(0)  # set NaNs to 0
success_df = success_df.drop(columns=["interval"])
success_df

study_intervals = [(0, 2000), (2000, 4000), (5000, 7000), (7000, 9000), (9000, 11000)]
print(f"{'INTERVAL':14} | {'LAUNCHES':10} | {'Success %':10} | {'Failed %':10} |")
print("-" * (14 + 10 + 10 + 10 + 11))
for interval in study_intervals:
    min_, max_ = interval
    cond = (success_df["min"] >= min_) & (success_df["max"] <= max_)
    temp_df = success_df.loc[cond]
    launches = temp_df["launches"].sum()
    succeeded = temp_df["success"].sum()
    failed = temp_df["failure"].sum()
    succeeded_pct = 0
    failed_pct = 0
    if launches > 0:
        succeeded_pct = round(succeeded / launches, 2)
        failed_pct = round(failed / launches, 2)

    print(
        f"{str(interval):14} | {launches:10d} | {succeeded_pct:10.2f} | {failed_pct:10.2f}"
    )

## 5. Booster with highest number of launches


# add new column to host booster versions
df["version"] = df["booster_version"].apply(lambda x: x.split()[1])

# intermediary lists
launches = []
succeeded = []
failed = []

# populate intermediary lists
for booster_version in unique_booster_versions:
    launches.append(df.loc[(df["version"] == booster_version)]["class"].count())
    succeeded.append(
        df.loc[(df["version"] == booster_version) & (df["class"] == 1)]["class"].count()
    )
    failed.append(
        df.loc[(df["version"] == booster_version) & (df["class"] == 0)]["class"].count()
    )

# create a new DataFrame to host booster versions, succeeded and failed launches
success_df = pd.DataFrame(
    data={
        "booster": unique_booster_versions,
        "launches": launches,
        "success": succeeded,
        "failure": failed,
    }
)
success_df["success_ratio"] = round(success_df["success"] / success_df["launches"], 4)
success_df["failure_ratio"] = round(success_df["failure"] / success_df["launches"], 4)
success_df.sort_values(by="success_ratio", ascending=False)
